requirements.txt
# dependencies
peewee>=3.14.4,<3.15.0;
python-dotenv>=0.18.0,<0.19.0;
fastapi>=0.65.2,<0.66.0;
uvicorn>=0.14.0,<0.15.0;
requests>=2.25.1,<2.26.0;
pyhumps>=3.0.2,<3.1.0;

# dev-dependencies
pytest>=6.2.4,<6.3.0;
flake8>=3.9.2,<3.10.0;
